### 2018-04-21 #687 @arfon
@gregreen - please merge this PR to fix your paper metadata https://github.com/gregreen/dustmaps/pull/10
