### 2018-04-23 #695 @whedon
[REVIEW]: dustmaps: A Python interface for maps of interstellar dust
**Submitting author:** @gregreen (<a href="http://orcid.org/0000-0001-5417-2260">Gregory Green</a>)
**Repository:** <a href="https://github.com/gregreen/dustmaps" target ="_blank">https://github.com/gregreen/dustmaps</a>
**Version:** v1.0
