### 2018-04-23 #695 @gregreen
@arfon Is there an equivalent to `\citet{}` in mardown? Some of the citations should not be in parentheses.
