### 2018-04-23 0de66ff
Fixed citation style in paper.
