### 2018-04-23 1dde6ab
Fixed figure label in paper.
