### 2018-06-05 #695 @jegpeek
Yep — been traveling with crappy wifi so hard to kick all the tires. Will do!
