### 2018-06-05 #695 @gregreen
I would be happy to do so. I can't access http://peekandgraves2010.com/ though. Is there a newer link to the data? Better yet, is the data stored in some semi-permanent archive (like the Harvard Dataverse)? I write a `fetch()` function for each dust map, so having a stable URL is best.
