### 2018-06-11 #695 @gregreen
I've added an issue, which I'll work on, but don't let that get in the way of looking over the code as it already is!
