### 2018-06-14 #695 @gregreen
@jegpeek Have you had time to test the functionality of the package, or to look at the documentation?
