### 2018-06-15 #695 @jegpeek
@gregreen making progress! Right now I am having this issue: I was following along in the directions [here](http://dustmaps.readthedocs.io/en/latest/examples.html#querying-reddening-at-an-array-of-coordinates):

```
import dustmaps.planck
dustmaps.planck.fetch()
```
and then got

```
Downloading http://pla.esac.esa.int/pla/aio/product-action?MAP.MAP_ID=HFI_CompMap_ThermalDustModel_2048_R1.20.fits ...
KeyError                                  Traceback (most recent call last)
<ipython-input-10-3c6fb0816ca2> in <module>()
      1 import dustmaps.planck
----> 2 dustmaps.planck.fetch()

/Users/jegpeek/anaconda3/lib/python3.5/site-packages/dustmaps/planck.py in fetch()
    130         'planck',
    131         'HFI_CompMap_ThermalDustModel_2048_R1.20.fits')
--> 132     fetch_utils.download_and_verify(url, md5, fname=fname)
    133
    134

/Users/jegpeek/anaconda3/lib/python3.5/site-packages/dustmaps/fetch_utils.py in download_and_verify(url, md5sum, fname, chunk_size, clobber, verbose)
    234
    235             with open(fname, 'wb') as f:
--> 236                 content_length = int(r.headers['content-length'])
    237                 bar = FileTransferProgressBar(content_length)
    238

/Users/jegpeek/anaconda3/lib/python3.5/site-packages/requests/structures.py in __getitem__(self, key)
     52
     53     def __getitem__(self, key):
---> 54         return self._store[key.lower()][1]
     55
     56     def __delitem__(self, key):

KeyError: 'content-length'
```

any ideas?
