### 2018-06-15 #695 @gregreen
It looks like the website isn't returning "content-length" in the headers. I'll fix the case where "content-length" is missing.
