### 2018-06-15 #695 @jegpeek
<KHAAAN.gif, but instead, PLAAAANCK>
also
<and_I_helped.gif>
