### 2018-06-15 #695 @gregreen
@jegpeek Can you verify that latest commit in the develop branch (gregreen/dustmaps@4c48e989c893a76f8107bf28d3e2620fd704b6ae) fixes the "content-length" issue?
