### 2018-06-24 #695 @jegpeek
Hey @gregreen I'm a little worried about the scaling issue, especially for bayestar. It's kinda confusing to report things "in SFD units" by default. There are two corrections discussed in GSF, I think (?) and I have trouble understanding which are which, and which are applied. This
```
sfd = SFDQuery()
Av_sfd = 2.742 * sfd(coords)

planck = PlanckQuery()
Av_planck = 3.1 * planck(coords)

bayestar = BayestarQuery(max_samples=1)
Av_bayestar = 2.742 * bayestar(coords)
```
makes me think bayestar is reported in "SFD units". I haven't met anyone outside the three of you who finds this natural (not that I have taken a poll...) . I get that the map is reported in these units, and thus that ship has in some sense sailed. Could there be either clearer documentation on this, or a correction=False flag, or something? I'll bet a ton of people are applying this naively.
