### 2018-06-27 016eafd
Fixing issue #11
