### 2018-06-27 #695 @jegpeek
@gregreen I know you guys have thought about this a lot, but it might be worth reconsidering the default behavior for these functions regarding "EBV_SFD". I don't think it's my role here as JOSS reviewer to get into it, so I'm not going to hold up the process or anything, but I think if you are going to all the trouble to package all of this so elegantly, asking folks to think in unrescaled SFD units and then multiplying is pretty weird. Maybe the package
