### 2018-06-27 #695 @gregreen
@jegpeek There are tests [here](https://github.com/gregreen/dustmaps/tree/master/dustmaps/tests). You can go to the dustmaps/ directory in the repository and run

    nosetests -v

Be careful though - some of the tests use a lot of memory! If you want to reduce the memory usage of the Bayestar map test, you can change [this line](https://github.com/gregreen/dustmaps/blob/master/dustmaps/tests/test_bayestar.py#L106) to `max_samples = 1`.
