### 2018-06-27 #695 @gregreen
@jegpeek No problem. Thanks for looking through my code/docs/paper!
