### 2018-06-27 f575a3a
Fixing JSON serialization of strings in Python3.
