+++ docs/conf.py
@@ -68,9 +68,9 @@ author = u'Gregory M. Green'
 # built documents.
 #
 # The short X.Y version.
-version = u'v1.0'
+version = u'v1.0.1'
 # The full version, including alpha/beta/rc tags.
-release = u'v1.0'
+release = u'v1.0.1'
 
 # The language for content autogenerated by Sphinx. Refer to documentation
 # for a list of supported languages.
