### 2018-06-27 10f44db
Update README.md
Added Zenodo DOI.
